Sp: Remove old sphinx docs

+++ sp/users/intents.py
@@ -28,6 +28,7 @@ DEFAULT_INTENT_NAME = "main"
 # Вспомогательные классы
 # ======================
 
+
 class IntentObject(NamedTuple):
     """Описывает одно намерение пользователя в хранилище.
 
@@ -35,11 +36,6 @@ class IntentObject(NamedTuple):
     Это позволяет легко получат намерения по их названию.
     Также по данному ключу может и не быть намерения, или его
     невозможно будет прочитать.
-
-    :param name: Имя намерения, используется как ключ.
-    :type name: str
-    :param intent: Экземпляр пользовательского намерения.
-    :type intent: Intent | None
     """
 
     name: str
@@ -48,6 +44,7 @@ class IntentObject(NamedTuple):
 
 # Основной класс хранилища
 # ========================
+
 
 class UserIntentsStorage:
     """Хранилище пользовательских намерений для платформы.
@@ -61,11 +58,6 @@ class UserIntentsStorage:
     Хранилище построено на базе данных Sqlite, что вероятно не самый
     оптимальный выбор, учитывая что данные хранятся в формате
     ключ - значение.
-
-    :param path: Путь к файлу базы данных для конкретной платформы.
-    :type path: str
-    :param uid: ID пользователя расписание.
-    :type uid: str
     """
 
     def __init__(self, path: str | Path, uid: int) -> None:
@@ -75,13 +67,13 @@ class UserIntentsStorage:
         self._check_tables()
 
     def _check_tables(self) -> None:
-        self._db.execute("CREATE TABLE IF NOT EXISTS intent("
+        self._db.execute(
+            "CREATE TABLE IF NOT EXISTS intent("
             "user_id TEXT NOT NULL,"
             "name TEXT NOT NULL,"
             "intent TEXT NOT NULL)"
         )
         self._db.commit()
-
 
     # Получение намерений
     # ===================
@@ -90,17 +82,12 @@ class UserIntentsStorage:
         """Получает количество намерений пользователя."""
         cur = self._db.cursor()
         cur.execute(
-            "SELECT name,intent FROM intent WHERE user_id=?",
-            (self.uid,)
+            "SELECT name,intent FROM intent WHERE user_id=?", (self.uid,)
         )
         return len(cur.fetchall())
 
     def __iter__(self) -> Iterator[IntentObject]:
-        """Позволяет по очереди получать каждое намерение пользователя.
-
-        :return: Экземпляр намерения из базы данных.
-        :rtype: IntentObject
-        """
+        """Позволяет по очереди получать каждое намерение пользователя."""
         return self.get()
 
     def _get_intent_object(self, name: str, intent_str: str) -> IntentObject:
@@ -108,8 +95,7 @@ class UserIntentsStorage:
             intent = Intent.from_str(intent_str)
         except Exception as e:
             logger.error(
-                "Error while unpack intent {} ({}): {}",
-                name, intent_str, e
+                "Error while unpack intent {} ({}): {}", name, intent_str, e
             )
             intent = Intent()
 
@@ -119,15 +105,11 @@ class UserIntentsStorage:
         """Получает список всех намерений пользователя.
 
         Намерения возвращаются в виде списка IntentObject.
-
-        :yield: Экземпляр намерения пользователя.
-        :rtype: list[IntentObject]
         """
         # Получаем имя и намерение из списка намерений для пользователя.
         cur = self._db.cursor()
         cur.execute(
-            "SELECT name,intent FROM intent WHERE user_id=?",
-            (self.uid,)
+            "SELECT name,intent FROM intent WHERE user_id=?", (self.uid,)
         )
         for n, i in cur:
             yield self._get_intent_object(n, i)
@@ -140,16 +122,11 @@ class UserIntentsStorage:
         Если есть несколько намерений с одинаковым именем - вернёт
         первое.
         Если намерений с таким именем нет, то вернёт None.
-
-        :param name: Имя намерения для поиска.
-        :type name: str
-        :return: Экземпляр намерения или None, если нечего на найдено.
-        :rtype: Optional[Intent]
         """
         cur = self._db.cursor()
         cur.execute(
             "SELECT intent FROM intent WHERE user_id=? AND name=?",
-            (self.uid, name)
+            (self.uid, name),
         )
         res = cur.fetchone()
         if res is not None:
@@ -165,7 +142,6 @@ class UserIntentsStorage:
         self._db.execute("DELETE FROM intent WHERE user_id=?", (self.uid,))
         self._db.commit()
 
-
     # Управление конкретным намерением
     # ================================
 
@@ -175,23 +151,18 @@ class UserIntentsStorage:
         Вы можете использовать этот метод как для создания нового
         намерение, так и для изменения старого.
         Если такое намерение уже существует, то будет перезаписано.
-
-        :param name: Имя намерения для добавления/изменения.
-        :type name: str
-        :param intent: Экземпляр намерения для добавления в базу данных.
-        :type intent: Intent
         """
         int_s = intent.to_str()
         cur = self._db.cursor()
         if self.get_intent(name) is not None:
             cur.execute(
                 "UPDATE intent SET intent=? WHERE user_id=? AND name=?",
-                (int_s, self.uid, name)
+                (int_s, self.uid, name),
             )
         else:
             cur.execute(
                 "INSERT INTO intent(user_id,name,intent) VALUES(?,?,?);",
-                (self.uid, name, int_s)
+                (self.uid, name, int_s),
             )
         self._db.commit()
 
@@ -203,15 +174,10 @@ class UserIntentsStorage:
         Обратите внимание, не происходит никаких проверок.
         Если такого намерения не существует - ничего не произойдёт.
         Статусных кодов возврата также не предусмотрено.
-
-        :param old_name: Старое имя намерения.
-        :type old_name: str
-        :param new_name: Новое имя намерения.
-        :type new_name: str
         """
         self._db.execute(
             "UPDATE intent SET name=? WHERE user_id=? AND name=?",
-            (new_name, self.uid, old_name)
+            (new_name, self.uid, old_name),
         )
         self._db.commit()
 
@@ -221,16 +187,11 @@ class UserIntentsStorage:
         Обратите внимание что у этого метода нет проверок.
         Если вы укажете не существующее имя, то ничего не произойдёт.
         Статусных кодов также не предусмотрено.
-
-        :param name: Имя намерения для удаления из базы данных.
-        :typw name: str
         """
         self._db.execute(
-            "DELETE FROM intent WHERE user_id=? AND name=?",
-            (self.uid, name)
+            "DELETE FROM intent WHERE user_id=? AND name=?", (self.uid, name)
         )
         self._db.commit()
-
 
     # Намерение по умолчанию
     # ======================
@@ -243,9 +204,6 @@ class UserIntentsStorage:
         Используется когда нам нужно передавать в классы представления
         некоторое намерение, однако нам его не передали.
         Тогда в ход вступает намерение по умолчанию.
-
-        :return: Намерение по умолчанию или пустое намерение намерение.
-        :rtype: Intent
         """
         intent = self.get_intent(name=DEFAULT_INTENT_NAME)
         return intent if intent is not None else Intent()
@@ -259,9 +217,6 @@ class UserIntentsStorage:
         Это будет полезно в тех случаях, когда мы хотим узнать,
         устанавливал пользователь намерение по умолчанию или нет.
         В некотором роде для обратной совместимости.
-
-        :return: Намерение пользователя по умолчанию.
-        :rtype: Optional[Intent]
         """
         return self.get_intent(name=DEFAULT_INTENT_NAME)
 
@@ -272,8 +227,5 @@ class UserIntentsStorage:
         методом мы задаём намерение пользователя по умолчанию.
         Оно будет использовать для автоматической подстановки в методы
         класса представления.
-
-        :param intent: Устанавливаемое намерение по умолчанию.
-        :type intent: Intent
         """
         self.add(name=DEFAULT_INTENT_NAME, intent=intent)
